+++ src/main/java/tahub/sdapitahub/controller/RoleController.java
@@ -5,8 +5,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import tahub.sdapitahub.constants.RoleMsgs;
 import tahub.sdapitahub.dto.Role.RoleDTO;
 import tahub.sdapitahub.dto.Role.RoleGetDTO;
+import tahub.sdapitahub.entity.Client;
 import tahub.sdapitahub.entity.Role;
 import tahub.sdapitahub.service.RoleService;
 
@@ -38,24 +40,42 @@ public class RoleController {
     }
 
     @PostMapping("/role")
-    public ResponseEntity<Role> createRole(@RequestBody RoleDTO roleCreateDTO) {
-        Role createdRole = roleService.createRole(roleCreateDTO);
-        return new ResponseEntity<>(createdRole, HttpStatus.CREATED);
-    }
-
-    @PutMapping("/role/{id}")
-    public ResponseEntity<Role> updateRole(@PathVariable("id") Long id, @RequestBody RoleDTO roleCreateDTO) {
-        Role updatedRole = roleService.updateRole(id, roleCreateDTO);
-        if (updatedRole != null) {
-            return new ResponseEntity<>(updatedRole, HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    public ResponseEntity<String> createRole(@RequestBody RoleDTO roleCreateDTO) {
+        try {
+            Role createdRole = roleService.createRole(roleCreateDTO);
+            return new ResponseEntity<String>(RoleMsgs.ROLE_CREATED.getMessage(), HttpStatus.CREATED);
+        } catch (Exception e) {
+            return new ResponseEntity<String>(RoleMsgs.ERROR_CREATE.getMessage(),HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
+    @PutMapping("/role/{id}")
+    public ResponseEntity<String> updateRole(@PathVariable("id") Long id, @RequestBody RoleDTO roleCreateDTO) {
+        try {
+            Role updatedRole = roleService.updateRole(id, roleCreateDTO);
+            if (updatedRole != null) {
+                return new ResponseEntity<String>(RoleMsgs.ROLE_UPDATED.getMessage(), HttpStatus.OK);
+            } else {
+                return new ResponseEntity<String>(RoleMsgs.ERROR_UPDATE.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+        } catch (Exception e) {
+            return new ResponseEntity<String>(RoleMsgs.ROLE_NOT_FOUND.getMessage(), HttpStatus.NOT_FOUND);
+
+        }
+    }
     @DeleteMapping("/role/{id}")
-    public ResponseEntity<Void> deleteRole(@PathVariable("id") Long id) {
-        roleService.deleteRole(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    public ResponseEntity<String> deleteRole(@PathVariable("id") Long id) {
+     try{
+         Optional<Role>existingRole = roleService.getRoleById(id);
+         if(existingRole.isPresent()){
+             roleService.deleteRole(id);
+             return new ResponseEntity<String>(RoleMsgs.ROLE_DELETED.getMessage(), HttpStatus.OK);
+         }else{
+           return new ResponseEntity<String>(RoleMsgs.ERROR_DELETE.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+         }
+     }catch (Exception e) {
+         return new ResponseEntity<String>(RoleMsgs.ROLE_NOT_FOUND.getMessage(), HttpStatus.NOT_FOUND);
+        }
+
     }
 }
